NIM-11 Run test coverage and send the results to codecov.io (#11)
Not running `codecov` in `tox` because you need a token to upload the results,
while you don't need one when it's running on Travis. Note though that running
`tox` will show the test coverage, so you have the data, it just won't upload
it.

Had to rename the environments in `tox` because `tox-travis` looks for
environments based on the python versions we give it. In this case, in
`.travis.yml`, we specified `"3.6"` so it'll look for environments that contain
`py36`. It makes `tox.ini` a bit uglier but on the bright side, now Travis
actually runs our tests!