NIM-8 Test code in the README
Install `pytest-readme` to test the code we put in the `README`,
whenever we get to add some code to the README that is.

`pytest-readme` requires pytest to be run with `--doctest-modules`, so I
added that option. Now `pytest` will also check the code in our
docstrings, again whenever we add some, which is cool.

When running, it will create a new file called `test_readme.py` so
adding it to .gitignore.

Passing the name of the app to pytest because otherwise the
`--doctest-modules` flag makes it hang. Could be because it's testing
the docstrings of code inside the `.tox` directory.

Created `conftest.py` as this is required for `pytest-readme` to work.