Add classifiers to setup.py (#24)
This may fix the issue that landscape.io is having when testing our app.

The issue could be https://github.com/landscapeio/landscape-issues/issues/300
and one of the reason it could be failing is that if we don't specify
Python 3, landscape.io will assume Python 2:
https://github.com/landscapeio/landscape-docs/blob/master/configuration.rst#python-targets-2-3

One way to fix that is to add classifiers to setup.py to explicitely say
we need Python 3. And while I was at it, I also added classifiers that I
thought were relevant. The list of available classifiers can be found at
https://pypi.python.org/pypi?%3Aaction=list_classifiers