NIM-8 Configure pytest (#20)
* NIM-8 Configure pytest

- Run the tests in a random order to reveal unwanted coupling between
  tests using the pytest-random-order package.
- Simplify our tests using @pytest.mark.parametrize
- Make pylint happier by disabling a few errors

* NIM-8 Prettier test output

* NIM-8 Verbose output for @pytest.mark.parametrize

* NIM-8 Find commented out code

* NIM-8 Simplify linting and add pep257

* NIM-8 Use eradicate instead of pytest-eradicate

Commented out code is more about style than an actual test, so have
eradicate be run as part of the lint environment.

* NIM-8 Test code in the README

Install `pytest-readme` to test the code we put in the `README`,
whenever we get to add some code to the README that is.

`pytest-readme` requires pytest to be run with `--doctest-modules`, so I
added that option. Now `pytest` will also check the code in our
docstrings, again whenever we add some, which is cool.

When running, it will create a new file called `test_readme.py` so
adding it to .gitignore.

Passing the name of the app to pytest because otherwise the
`--doctest-modules` flag makes it hang. Could be because it's testing
the docstrings of code inside the `.tox` directory.

Created `conftest.py` as this is required for `pytest-readme` to work.

* NIM-8 Remove stale bytecode before running tests

* NIM-8 Fold captured output on Travis

* NIM-8 Remove test_readme.py so eradicate doesn't run on it

* NIM-8 Clean up formatting

* NIM-8 Create security environment and add pyt